refactor: split audit.rs into focused submodules (#112 PR 1)
Split the monolithic audit.rs (2,811 lines) into a directory module
with 5 focused files:

- audit/mod.rs:       AuditConfig, AuditLogger, AuditEvent + all tests
- audit/chain.rs:     HMAC hash chain (HashableEvent, genesis, compute)
- audit/retention.rs: Prune logic (try_prune, build_prune_point)
- audit/secret.rs:    Secret management, flock, O_NOFOLLOW, key rotation
- audit/verify.rs:    Chain verification, show_entries, audit_summary

Visibility strategy per security plan:
- Public API unchanged (omamori::audit::* paths preserved via re-exports)
- Internal functions use pub(super) only (HMAC/secret primitives)
- build_prune_point and display_timestamp promoted to pub(super) for tests

Additional fixes discovered during split:
- read_events test helper: use filter_map to skip torn lines gracefully
- create_event_all_fields: fix expected result "blocked" → "block"
- create_event_all_fields: fix expected target_count 2 → 1 (--force filtered)
- RuleConfig test construction: update to current field names
